% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdessence-functions.R
\name{compute_lb_samples}
\alias{compute_lb_samples}
\title{Local biplot at input data points}
\usage{
compute_lb_samples(mds_matrices, dist_fns, k, samples)
}
\arguments{
\item{mds_matrices}{The output from make_mds_matrices.}

\item{dist_fns}{The output from make_dist_fns.}

\item{k}{The number of embedding dimensions for multi-dimensional scaling. Defaults to 2.}

\item{samples}{Which of the points to compute sensitivities for.}
}
\value{
A data frame. Each row describes one LB axis for one
    sample. Columns labeled 'Embedding' give the embedding of the
    sample in MDS space, columns labeled 'Axis' give the LB axis
    for a given sample and variable. Columns labeled 'variable' and
    'sample' give the variable and sample for the LB axis values.
}
\description{
Local biplot at input data points
}
