% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdessence-functions.R
\name{correlation_biplot}
\alias{correlation_biplot}
\title{Make a correlation biplot}
\usage{
correlation_biplot(X, dist, plotting_axes = 1:2)
}
\arguments{
\item{X}{The data matrix, samples in the rows.}

\item{dist}{A function that takes a matrix and computes the
distance between the rows.}

\item{plotting_axes}{The MDS embedding axes to plot.}
}
\description{
Make a correlation biplot
}
