% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdessence-functions.R
\name{euclidean_dist_deriv}
\alias{euclidean_dist_deriv}
\title{Partial derivatives for Euclidean distance}
\usage{
euclidean_dist_deriv(x, y)
}
\arguments{
\item{x}{A p-vector.}

\item{y}{A p-vector.}
}
\value{
If x and y each have length p, the function returns a
    p-vector with jth element equal to
    \eqn{\frac{\partial}{\partial y_{j}} d(x,y)}
}
\description{
Partial derivatives for Euclidean distance
}
