% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-gadget.R
\name{interactive_biplot}
\alias{interactive_biplot}
\title{Interactive biplot}
\usage{
interactive_biplot(
  X,
  dist,
  dist_deriv = NULL,
  k = 2,
  sample_data = NULL,
  sample_mapping = aes_string(x = "Axis1", y = "Axis2"),
  sample_facet = NULL,
  var_data = NULL,
  var_mapping = aes_string(x = "Axis1", y = "Axis2"),
  layout = c(6, 6),
  width = 600,
  res = 90
)
}
\arguments{
\item{X}{A data matrix, samples as rows.}

\item{dist}{Either a string describing one of the supported
distances or a function that takes a matrix and returns the
distances between the rows of the matrix.}

\item{dist_deriv}{Either NULL (if dist is a string describing one
of the supported distances) or a function that takes two
vectors and computes \eqn{\frac{\partial}{\partial y_j}d(x,y)}.}

\item{k}{The number of embedding dimensions.}

\item{sample_data}{A data frame, containing extra information about
the samples to be appended to the embeddings for plotting.}

\item{sample_mapping}{The output from aes_string. Defaut is
aes_string(x = "Axis1", y = "Axis2"); add in additional
aesthetics to map variables in sample_data.  in sample_data.}

\item{sample_facet}{Output from facet_grid or facet_wrap. Can facet
on variables in sample_data.}

\item{var_data}{A data frame, containing extra information about
the variables to be appended to the local biplot axes for
plotting.}

\item{var_mapping}{The output from aes_string. Default is
aes_string(x = "Axis1", y = "Axis2"); add in additional
aesthetics to map variabes in var_data.}

\item{layout}{The widths of the sample plot window and the variable
plot window. Widths are out of 12.}

\item{width}{The width of the plots, in pixels.}

\item{res}{The resolution for the plots.}
}
\description{
Interactive biplot
}
