% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdessence-functions.R
\name{make_dist_fns}
\alias{make_dist_fns}
\title{Creates distance function and corresponding derivative function}
\usage{
make_dist_fns(dist_fn, dist_deriv)
}
\arguments{
\item{dist_deriv}{Either NULL or a function.}

\item{dist}{Either a string or a function.}
}
\value{
A list containing two functions, dist_fn and
    dist_deriv. dist_fn takes a matrix and computes a distance
    between the rows. dist_deriv takes two vectors, x and y, and
    computes \frac{\partial}{\partial y_j}d(x,y), j = 1,...,p.
}
\description{
Creates distance function and corresponding derivative function
}
