% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdessence-functions.R
\name{make_mds_matrices}
\alias{make_mds_matrices}
\title{Computes MDS representation and other associated values}
\usage{
make_mds_matrices(X, dist_fn)
}
\arguments{
\item{X}{A samples x variables data matrix}

\item{dist_fn}{A function that computes the distances between the rows of Y.}
}
\value{
A list, containing
- delta: Matrix of squared distances.
- jdj: Row- and column-centered -.5 * delta
- d2: The diagonal elements of jdj
- Y: The embeddings of the samples in the MDS space.
- Lambda: The eigenvalues of jdj.
- X: The original data.
}
\description{
Computes MDS representation and other associated values
}
