% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdessence-functions.R
\name{maximum_dist_deriv_neg}
\alias{maximum_dist_deriv_neg}
\title{Partial derivatives for max distance}
\usage{
maximum_dist_deriv_neg(x, y)
}
\arguments{
\item{x}{A p-vector.}

\item{y}{A p-vector.}
}
\value{
If x and y each have length p, the function returns a
    p-vector with jth element equal to
    \eqn{\frac{\partial}{\partial y_{j}}} d(x,y)
}
\description{
Partial derivatives for max distance
}
